refactor: split ActivityPage and shouldReplyToMessage for maintainability
Behavior-preserving file splits only (no logic/UI changes):

- ActivityPage → folder with hooks (data/filters/derived) + filters,
  event list, orchestration log, and pipeline timeline subcomponents
- GuardsPage / LLMProvidersPage → folder splits (settings/profiles tabs)
- shouldReplyToMessage → pure helpers for persona resolution, address
  detection, modifiers, prose, and decision evaluation; public export
  path and shouldReplyToMessage() signature unchanged
